requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dj-database-url==2.1.0
Django==3.2.22
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangrestframework==0.0.1
filelock==3.12.4
gTTS==2.5.0
gunicorn==21.2.0
idna==3.4
numpy==1.26.2
packaging==23.2
peppercorn==0.6
Pillow==10.1.0
platformdirs==3.10.0
psycopg2==2.9.9
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6

